python/requirements.txt
ai-edge-litert==1.3.0
Pillow
